Avoid bool vs int comparison warning while downcasting. JIRA: CXX-5700
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@66471 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objects/seqtable/SeqTable_multi_data.cpp
@@ -433,7 +433,7 @@ static inline
 bool sx_DownCastInt8(DstInt& v, const Int8& value, const char* type_name)
 {
     v = DstInt(value);
-    if ( v != value ) {
+    if ( Int8(v) != value ) {
         NCBI_THROW_FMT(CSeqTableException, eIncompatibleValueType,
                        "CSeqTable_multi_data::TryGet"<<type_name<<"(): "
                        "value is too big for requested type: "<<value);
